add: Check Internet Connection

+++ app/src/main/java/com/example/firewatch/domain/repositories/BurnRepositoryImpl.kt
@@ -1,6 +1,7 @@
 package com.example.firewatch.domain.repositories
 
-import com.example.firewatch.domain.daos.BurnDao
+import android.content.Context
+import com.example.firewatch.R
 import com.example.firewatch.domain.repositories.dtos.burn.BurnCreateInput
 import com.example.firewatch.domain.repositories.dtos.burn.BurnUpdateInput
 import com.example.firewatch.domain.entities.Burn
@@ -15,16 +16,28 @@ import com.example.firewatch.services.http.HttpService
 import com.example.firewatch.services.http.contracts.Pagination
 import com.example.firewatch.services.persistence.DatabaseContext
 import com.example.firewatch.shared.utils.DateUtils
-import java.math.BigDecimal
+import dagger.hilt.android.qualifiers.ApplicationContext
 import java.time.LocalDateTime
 
 class BurnRepositoryImpl(
+    @ApplicationContext private val context: Context,
     private val httpService: HttpService,
     private val connectivityService: ConnectivityService,
     private val dbContext: DatabaseContext
 ) : BurnRepository {
+    private fun updateState(id: String, state: BurnState) {
+        val burn = dbContext.burnDao().getById(id) ?: return
+        burn.state = state
+
+        dbContext.burnDao().update(burn)
+    }
+
     override suspend fun create(input: BurnCreateInput): Result<BurnRequest> {
         try {
+            if (!connectivityService.isConnectionActive()) {
+                throw Exception(context.getString(R.string.need_connection))
+            }
+
             val response = HttpService.fetch {
                 httpService.burnApiService.create(input.toMultipart())
             }
@@ -44,6 +57,10 @@ class BurnRepositoryImpl(
     }
 
     override suspend fun update(input: BurnUpdateInput): Result<Burn> = try {
+        if (!connectivityService.isConnectionActive()) {
+            throw Exception(context.getString(R.string.need_connection))
+        }
+
         val response = HttpService.fetch {
             httpService.burnApiService.update(input.id, input.toMultipart())
         }
@@ -51,12 +68,20 @@ class BurnRepositoryImpl(
         val result = response.getOrThrow()
         val burn = result.properties.toBurn(result.geometry.getCoordinate())
 
+        if (connectivityService.isConnectionActive()) {
+            dbContext.burnDao().update(burn)
+        }
+
         Result.success(burn)
     } catch (e: Exception) {
         Result.failure(e)
     }
 
     override suspend fun getAvailabitity(coordinates: Coordinates): Result<Boolean> = try {
+        if (!connectivityService.isConnectionActive()) {
+            throw Exception(context.getString(R.string.need_connection))
+        }
+
         val response = HttpService.fetch {
             httpService.burnApiService.getAvailability(
                 coordinates.lat,
@@ -72,6 +97,10 @@ class BurnRepositoryImpl(
     }
 
     override suspend fun terminate(id: String): Result<BurnState> = try {
+        if (!connectivityService.isConnectionActive()) {
+            throw Exception(context.getString(R.string.need_connection))
+        }
+
         val response = HttpService.fetch {
             httpService.burnApiService.terminate(
                 id
@@ -81,12 +110,17 @@ class BurnRepositoryImpl(
         val result = response.getOrThrow()
         val burnState = BurnState.get(result.state) ?: throw Exception("wasn't possible to convert to burn state")
 
+        updateState(result.burnId, burnState)
         Result.success(burnState)
     } catch (e: Exception) {
         Result.failure(e)
     }
 
     override suspend fun start(id: String): Result<BurnState> = try {
+        if (!connectivityService.isConnectionActive()) {
+            throw Exception(context.getString(R.string.need_connection))
+        }
+
         val response = HttpService.fetch {
             httpService.burnApiService.start(
                 id
@@ -95,6 +129,10 @@ class BurnRepositoryImpl(
 
         val result = response.getOrThrow()
         val burnState = BurnState.get(result.state) ?: throw Exception("wasn't possible to convert to burn state")
+
+        if (connectivityService.isConnectionActive()) {
+            updateState(result.burnId, burnState)
+        }
 
         Result.success(burnState)
     } catch (e: Exception) {
@@ -153,13 +191,17 @@ class BurnRepositoryImpl(
     ): Result<List<Burn>> {
         try {
             if (!connectivityService.isConnectionActive()) {
+                var searchPattern: String? = null
+
+                if (!search.isNullOrBlank()) {
+                    searchPattern = "%$search%"
+                }
+
                 val result = dbContext.burnDao().getAll(
-//                    search = search,
-//                    state = state,
-//                    startDate = startDate,
-//                    endDate = endDate,
-//                    limit = 10,
-//                    offset = 1
+                    search = searchPattern,
+                    state = state,
+                    startDate = startDate,
+                    endDate = endDate,
                 )
 
                 return Result.success(result)
@@ -188,27 +230,67 @@ class BurnRepositoryImpl(
         }
     }
 
-    override suspend fun get(id: String): Result<Burn> = try {
-        val response = HttpService.fetch {
-            httpService.burnApiService.getById(id)
+    override suspend fun get(id: String): Result<Burn> {
+        try {
+            if (!connectivityService.isConnectionActive()) {
+                val burn = dbContext.burnDao().getById(id) ?: throw Exception("burn not found")
+
+                return Result.success(burn)
+            }
+
+            val response = HttpService.fetch {
+                httpService.burnApiService.getById(id)
+            }
+
+            val result = response.getOrThrow()
+            val burn = result.properties.toBurn(result.geometry.getCoordinate())
+
+            return Result.success(burn)
+        } catch (e: Exception) {
+            return Result.failure(e)
         }
-
-        val result = response.getOrThrow()
-        val burn = result.properties.toBurn(result.geometry.getCoordinate())
-
-        Result.success(burn)
-    } catch (e: Exception) {
-        Result.failure(e)
     }
 
     override suspend fun delete(id: String): Result<String> = try {
+        if (!connectivityService.isConnectionActive()) {
+            throw Exception(context.getString(R.string.need_connection))
+        }
+
         val response = HttpService.fetch {
             httpService.burnApiService.delete(id)
         }
 
         val result = response.getOrThrow().burnId
+
+        if (connectivityService.isConnectionActive()) {
+            val burn = dbContext.burnDao().getById(id)
+
+            burn?.let {
+                dbContext.burnDao().delete(burn)
+            }
+        }
+
         Result.success(result)
     } catch (e: Exception) {
         Result.failure(e)
+    }
+
+    override suspend fun sync() {
+        if (!connectivityService.isConnectionActive()) {
+            return
+        }
+
+        val fetchBurns = getAll()
+
+        if (fetchBurns.isFailure) {
+            return
+        }
+
+        dbContext.burnDao().truncate()
+        val burns = fetchBurns.getOrThrow()
+
+        burns.forEach {
+            dbContext.burnDao().create(it)
+        }
     }
 }
